refactor: create abstract base class for backup tests
